requirements/production.txt
#Only production requirement for build
#=====================================
websocket-client==0.54.0